Add random race to matchup players.
Use the player's hero selection to determine, if possible,
which race they were assigned from the random selection.
Assigned to a new member property, so that it can be used
separately for statistical computations/display. If the
player's race is not random, then null is assigned to the
random race member. If it cannot be determined which race
was assigned, then random is set for both the player race
as well as the random race.