Add map filter to ongoing and previous matches.
Defaults to the "Overall" string, which matches the accompanying
change to the front-end. A future enhancement to this feature
would be an endpoint that could retrieve the map choices for a
given Gateway/GameMode combination. The front end could then
dynamically draw the map select from this data. This would
be a better user experience, since it would reduce the number
of empty result outcomes.