Fixing multiple clan issues
Setting all people in clan as "members" to solve an issue where shamans could join multiple clans.

Adding and fixing a few existing units.

Removing a link between founder and a member to allow founders not to be clan members

Creating a clan with non-unique name should not be possible (currently only id had to be unique)